move variables category blocks to common droplet config
